Extract out IPAddress and IUDPSocket
