+++ fdbrpc/sim2.actor.cpp
@@ -54,6 +54,7 @@
 #include "fdbrpc/AsyncFileWriteChecker.h"
 #include "flow/FaultInjection.h"
 #include "flow/TaskQueue.h"
+#include "flow/IUDPSocket.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 ISimulator* g_simulator = nullptr;
