+++ fdbserver/MetricLogger.actor.cpp
@@ -40,6 +40,7 @@
 #include "fdbserver/MetricClient.h"
 #include "flow/flow.h"
 #include "flow/network.h"
+#include "flow/IUDPSocket.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 struct MetricsRule {
