+++ fdbserver/workloads/UDPWorkload.actor.cpp
@@ -34,6 +34,7 @@
 #include <vector>
 #include <memory>
 #include <functional>
+#include "flow/IUDPSocket.h"
 
 #include "flow/actorcompiler.h" // has to be last include
