+++ flow/include/flow/network.h
@@ -22,15 +22,10 @@
 #define FLOW_OPENNETWORK_H
 #pragma once
 
-#include <array>
-#include <regex>
 #include <string>
 #include <stdint.h>
-#include <variant>
 #include <atomic>
-#include <boost/asio/ip/address_v6.hpp>
 #include <boost/asio/ip/tcp.hpp>
-#include <boost/asio/ip/udp.hpp>
 #include "flow/Arena.h"
 #include "flow/BooleanParam.h"
 #include "flow/IRandom.h"
@@ -38,72 +33,10 @@
 #include "flow/Trace.h"
 #include "flow/TaskPriority.h"
 #include "flow/WriteOnlySet.h"
+#include "flow/IPAddress.h"
 
 class Void;
 
-struct IPAddress {
-	typedef boost::asio::ip::address_v6::bytes_type IPAddressStore;
-	static_assert(std::is_same<IPAddressStore, std::array<uint8_t, 16>>::value,
-	              "IPAddressStore must be std::array<uint8_t, 16>");
-
-public:
-	// Represents both IPv4 and IPv6 address. For IPv4 addresses,
-	// only the first 32bits are relevant and rest are initialized to
-	// 0.
-	IPAddress() : addr(uint32_t(0)) {}
-	explicit IPAddress(const IPAddressStore& v6addr) : addr(v6addr) {}
-	explicit IPAddress(uint32_t v4addr) : addr(v4addr) {}
-
-	bool isV6() const { return std::holds_alternative<IPAddressStore>(addr); }
-	bool isV4() const { return !isV6(); }
-	bool isValid() const;
-
-	// Returns raw v4/v6 representation of address. Caller is responsible
-	// to call these functions safely.
-	uint32_t toV4() const { return std::get<uint32_t>(addr); }
-	const IPAddressStore& toV6() const { return std::get<IPAddressStore>(addr); }
-
-	std::string toString() const;
-	static Optional<IPAddress> parse(std::string const& str);
-
-	bool operator==(const IPAddress& addr) const;
-	bool operator!=(const IPAddress& addr) const;
-	bool operator<(const IPAddress& addr) const;
-
-	template <class Ar>
-	void serialize(Ar& ar) {
-		if constexpr (is_fb_function<Ar>) {
-			serializer(ar, addr);
-		} else {
-			if (Ar::isDeserializing) {
-				bool v6;
-				serializer(ar, v6);
-				if (v6) {
-					IPAddressStore store;
-					serializer(ar, store);
-					addr = store;
-				} else {
-					uint32_t res;
-					serializer(ar, res);
-					addr = res;
-				}
-			} else {
-				bool v6 = isV6();
-				serializer(ar, v6);
-				if (v6) {
-					auto res = toV6();
-					serializer(ar, res);
-				} else {
-					auto res = toV4();
-					serializer(ar, res);
-				}
-			}
-		}
-	}
-
-private:
-	std::variant<uint32_t, IPAddressStore> addr;
-};
 
 template <>
 struct Traceable<IPAddress> : std::true_type {
@@ -558,27 +491,6 @@ protected:
 	~INetwork() {} // Please don't try to delete through this interface!
 };
 
-class IUDPSocket {
-public:
-	//  see https://en.wikipedia.org/wiki/User_Datagram_Protocol - the max size of a UDP packet
-	// This is enforced in simulation
-	constexpr static size_t MAX_PACKET_SIZE = 65535;
-	virtual ~IUDPSocket();
-	virtual void addref() = 0;
-	virtual void delref() = 0;
-
-	virtual void close() = 0;
-	virtual Future<int> send(uint8_t const* begin, uint8_t const* end) = 0;
-	virtual Future<int> sendTo(uint8_t const* begin, uint8_t const* end, NetworkAddress const& peer) = 0;
-	virtual Future<int> receive(uint8_t* begin, uint8_t* end) = 0;
-	virtual Future<int> receiveFrom(uint8_t* begin, uint8_t* end, NetworkAddress* sender) = 0;
-	virtual void bind(NetworkAddress const& addr) = 0;
-
-	virtual UID getDebugID() const = 0;
-	virtual NetworkAddress localAddress() const = 0;
-	virtual boost::asio::ip::udp::socket::native_handle_type native_handle() = 0;
-};
-
 // DNSCache is a class maintaining a <hostname, vector<NetworkAddress>> mapping.
 class DNSCache {
 public:
@@ -599,6 +511,8 @@ public:
 private:
 	std::map<std::string, std::vector<NetworkAddress>> hostnameToAddresses;
 };
+
+class IUDPSocket;
 
 class INetworkConnections {
 public:
