+++ flow/network.cpp
@@ -18,10 +18,11 @@
  * limitations under the License.
  */
 
-#include "flow/Arena.h"
-#include "boost/asio.hpp"
+#include <boost/asio.hpp>
 
+#include "flow/Arena.h"
 #include "flow/network.h"
+#include "flow/IUDPSocket.h"
 #include "flow/flow.h"
 #include "flow/UnitTest.h"
